+++ CHANGELOG.md
@@ -9,6 +9,12 @@ Unreleased
 ### Changed
 * Add theme support
 
+### Removed
+* Removed toggle backgrounds tab and functionality from example loader
+
+### Added
+* Added CSS display tab to example loader
+
 6.24.0 - (June 16, 2020)
 ----------
 ### Changed